requirements.txt
# Jaysynth RAG Project - Requirements
# =====================================

# PDF Processing
pdfplumber==0.11.0
PyPDF2==3.0.1

# Embeddings
sentence-transformers==3.0.1

# Vector Store
chromadb==0.5.0
faiss-cpu==1.8.0

# LLM & Orchestration
openai==1.40.0
llama-index==0.11.0
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-openai==0.2.0
llama-index-vector-stores-chroma==0.2.0

# Utilities
python-dotenv==1.0.1
tiktoken==0.7.0
numpy==1.26.4
rich==13.7.1
